+++ ExampleMod/Projectiles/ExampleJavelinProjectile.png
