+++ ExampleMod/Projectiles/ExampleSpearProjectile.png
