+++ ExampleMod/Projectiles/ExampleYoyoProjectile.png
